refactor(examples/rag_chat): split specs into 1 file per specification
